+++ src/Sharpen.Engine/CSharpFeatures/AsynchronousStreams.cs
@@ -3,16 +3,16 @@ using static Sharpen.Engine.CSharpLanguageVersions;
 
 namespace Sharpen.Engine.CSharpFeatures
 {
-    public class AsyncStreams : ICSharpFeature
+    public class AsynchronousStreams : ICSharpFeature
     {
-        private AsyncStreams() { }
+        private AsynchronousStreams() { }
 
         public ImmutableArray<string> LanguageVersions { get; } = ImmutableArray.CreateRange(new[] { CSharp80 });
 
-        public string FriendlyName { get; } = "Async streams";
+        public string FriendlyName { get; } = "Asynchronous streams";
 
         public string LearnMoreUrl => "https://docs.microsoft.com/en-us/dotnet/csharp/whats-new/csharp-8#asynchronous-streams";
 
-        public static readonly AsyncStreams Instance = new AsyncStreams();
+        public static readonly AsynchronousStreams Instance = new AsynchronousStreams();
     }
 }