+++ CHANGELOG.md
@@ -3,10 +3,20 @@ All notable changes to the "Sharpen" extension will be documented in this file.
 
 The format of the file is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.10.1] - 2019-10-12
+### Fixed
+- Navigating with keyboard (up, down, left, right, ...) does not work in the Sharpen Results view (#29).
+- Selected text in the Sharpen Results view is unreadable when the view loses focus (#28).
+
+## [0.10.0] - 2019-10-10
 ### Added
+- "Consider using ??= operator instead of assigning result of the ?? operator" suggestion.
+- "Use ??= operator instead of assigning result of the ?? operator" suggestion.
 - "What is/was new in C# X.Y?" link to MSDN documentation on C# version in the Sharpen Results view.
 - "Learn more..." link to MSDN documentation on language features in the Sharpen Results view.
+
+### Changed
+- "Async streams" renamed to "Asynchronous streams".
 
 ## [0.9.0] - 2019-06-14
 ### Added